Added functionality to the edit button
I worked with the same Form that adds the books, tricking its behaviour a bit.
I created a new submit button each time the edit button is pressed and delete
it once has done its job. I works, although allows the user to erase all the
data (Which will make no sense for the user, but lets leave it like that)